# Optimize parser